chore: rename ApplicationActions
