+++ packages/frontend/src/lib/ApplicationActions.svelte
