+++ packages/frontend/src/lib/table/environment/ColumnActions.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-import EnvironmentActions from '../../EnvironmentActions.svelte';
+import ApplicationActions from '../../ApplicationActions.svelte';
 import type { ApplicationCell } from '../../../pages/applications';
 export let object: ApplicationCell;
 </script>
 
-<EnvironmentActions recipeId="{object.recipeId}" modelId="{object.modelId}" object="{object.envState}" />
+<ApplicationActions recipeId="{object.recipeId}" modelId="{object.modelId}" object="{object.envState}" />
